Add support for different backends
Also implements ffmpeg as a new backend.

While a lot of changes were made in the
encoding process, this should produce the
exact same commands given the user does not
swap out backends (with mpv obviously being
the default option added).

Support for ffmpeg for everything is also
not 100% guaranteed, but it covers the
basic stuff.