encode: display encode progress
This creates a new class, EncodeWithProgress, that handles reading
mpv's output, and displaying the encode progress on the UI. The
code is a bit clunky, and there are a few 'hacks' that I wouldn't
like to use:

- because mpv detects whether we're outputting to a tty (terminal)
or not, and, when outputting to a terminal, it uses CR for the
status line, we need to write to a temporary file to be able to
catch the statusline updates. Maybe there's a way to read up to a
CR with Lua, but I wasn't able to find it.

- also, due to this, when waiting for the next line, it busy-waits
until the next line is available. This shouldn't be that big of a
issue, but... it still annoys me. Using timeouts could work too,
but it would make the code a bit more convoluted (even more than it
already is)

Maybe using mpv's IPC protocol could be a better way to work around
this, but, for now, this works.