encode: patch libvpx 2-pass log file
The pass log file generated by mpv doesn't contain the correct duration
of each frame/aggregated sum, which leads to garbage encodes
when used on the second pass.

Workaround this by writing the average frame duration on the frame
stats, and the total encode duration on the EOS packet. This should be
fairly accurate when dealing with constant frame rate videos, which,
hopefully, is the most common case.