vp8_twopass_log_patcher, util: change how files are read/written on windows
oh boy this is a mess

So, Lua can't read files with Unicode characters on the path at all on
Windows, and we need this funny workaround here. I'm really afraid
this will break very very soon, but for now let's keep this around.

Fixes #81 (but not really, I hope someday this isn't needed anymore).