Use better quality defaults (#130)
* formats/webm: use better quality defaults

Those mostly work only on twopass, but I don't think there's a problem
using them even if twopass is disabled

* options: enable twopass by default

* tests/mpv_ipc: fix possible race condition

When we send the request, it's possible that we get a reply before we
start waiting for the response, in which case our threading.Event
object can no longer be obtained through the dict.

* formats/webm: disable twopass support on VP9

It seems to be just as broken as VP8; and we don't really patch the pass
logfile for it yet. Maybe someday?