Add target file size to replace table
I write the file size to the name of the encoded file because different platforms have different upload limits. I often encode different files and differentiate the names by file size so that they don't overwrite. I do this by adding the following line to webm.lua. It would be nice if this was part of the official script so that I don't need to edit it each time I update.