Add options.target_filesize to replace table
This commit adds options.target_filesize to the replace table for writing file names. This is a change I always add when updating webm.lua. This is added because different platforms have different file size upload limits, so I often encode the same clip at different file sizes. The file names are then differentiated by size so that they don't overwrite. I'm submitting a pull request for this so that I don't need to re-add this to the script every time I update it.